resolve pull request
