+++ components/Providers.tsx
@@ -1,11 +1,8 @@
 "use client";
 
+import { PropsWithChildren } from "react";
 import { ParallaxProvider } from "react-scroll-parallax";
 
-export function Providers({
-  children,
-}: {
-  readonly children: React.ReactNode;
-}) {
+export function Providers({ children }: PropsWithChildren<unknown>) {
   return <ParallaxProvider>{children}</ParallaxProvider>;
 }
