+++ lib/config/page.config.ts
@@ -1,5 +1,5 @@
-import { GraphicProps, graphicConfig } from "@components/page/Graphic";
 import { FlexProps, flexConfig } from "@components/puck/Flex";
+import { GraphicProps, graphicConfig } from "@components/puck/Graphic";
 import { HeadingProps, headingConfig } from "@components/puck/Heading";
 import { HeroProps, heroConfig } from "@components/puck/Hero";
 import { IFrameProps, iframeConfig } from "@components/puck/IFrame";
@@ -18,13 +18,13 @@ import type { Config, Data } from "@measured/puck";
 // @keep-sorted
 export type PageProps = {
   Flex: FlexProps;
+  Graphic: GraphicProps;
   Heading: HeadingProps;
   Hero: HeroProps;
   IFrame: IFrameProps;
   SectionDivider: SectionDividerProps;
   Text: TextProps;
   VerticalSpace: VerticalSpaceProps;
-  Graphic: GraphicProps;
 };
 export type PageRootProps = {
   title: string;
@@ -36,13 +36,13 @@ export const pageConfig: PageConfig = sectionThemedConfig({
   // @keep-sorted
   components: {
     Flex: flexConfig,
+    Graphic: graphicConfig,
     Heading: headingConfig,
     Hero: heroConfig,
     IFrame: iframeConfig,
     SectionDivider: sectionDividerConfig,
     Text: textConfig,
     VerticalSpace: verticalSpaceConfig,
-    Graphic: graphicConfig,
   },
 });
